backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
SQLAlchemy
psycopg2-binary
black
python-jose[cryptography]
bcrypt==4.0.1             
passlib[bcrypt]
alembic
loguru
httpx